Fix BASE_5_STAR_COUNT type from double to int for consistency

+++ cpp/artifact/constants.hpp
@@ -9,7 +9,7 @@
 namespace Distribution {
 
 constexpr int RESIN_PER_RUN = 20;
-constexpr double BASE_5_STAR_COUNT = 1;            // One 5* Artifact per run
+constexpr int BASE_5_STAR_COUNT = 1;            // One 5* Artifact per run
 constexpr double DOUBLE_5_STAR_CHANCE = 0.065;     // 6.5% Chance to get a second 5* Artifact per run
 constexpr double SET_SPLIT_RATE = 0.50;            // Two sets per domain are equally split
 constexpr double SLOT_SPLIT_RATE = 0.20;           // Each Artifact Slot has an equal likelihood (20%) of dropping
